+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to the "html-translator" extension will be documented in thi
 * Support for custom escaping strings
 * Completition matching does not use regexp anymore
 * Multiple escapes on same line
+* Snippet to close opening escape string
 
 ## 0.0.2